modified:   src/handlers/handlers.py 	new file:   src/handlers/handlers_address.py 	new file:   src/handlers/handlers_birthday.py 	new file:   src/handlers/handlers_contact.py 	new file:   src/handlers/handlers_email.py 	new file:   src/handlers/handlers_note.py 	new file:   src/handlers/handlers_tags.py
